labels/scripts/requirements.txt
# Core Data Processing
numpy
pandas
scipy

# Database Connectivity
sqlalchemy
psycopg2-binary
geoalchemy2

# Machine Learning
scikit-learn
threadpoolctl # Required for scikit-learn
sentence-transformers

# Geospatial Processing
geopandas
shapely
pyproj  # Required for geopandas
fiona  # Required for geopandas

# Visualization
matplotlib
seaborn
folium
branca  # Required for folium
networkx

# Jupyter (optional for exploration)
jupyter
ipython

# Utilities
python-dotenv  # For environment variables
tqdm  # For progress bars

# Testing
pytest
pytest-cov  # For coverage reports
# Logging
logging  # Standard library, no need to install

# AI
google-generativeai #For Gemini


